Update README.md
This project applies Multiple Linear Regression to predict the profit of startups using the 50_Startups.csv dataset. The dataset includes key features such as R&D Spend, Administration, Marketing Spend, and State. The goal is to build a model that learns from these inputs to estimate a startup's potential profit.